requirements.txt
langchain==0.3.25
langchain-openai==0.3.24
openai-whisper==20240930
yt-dlp==2025.6.9 